RavenDB-25216 test(ui): share import-from-file DTO fixtures between specs
