+++ src/Raven.Studio/typescript/components/pages/database/tasks/importData/importFromFile/ImportDatabaseFromFile.spec.tsx
@@ -5,9 +5,18 @@ import { act, rtlRender, waitFor } from "test/rtlTestUtils";
 import { mockServices } from "test/mocks/services/MockServices";
 import { mockStore } from "test/mocks/store/MockStore";
 import { defaultTransformScript } from "./useImportFromFileForm";
+import {
+    adminDefaultDto,
+    adminDefaultOperateOnTypes,
+    allConnectionStringTokens,
+    allOngoingTaskTokens,
+    allSettingTokens,
+    operateOnTypesOf,
+    recordTypesOf,
+    without,
+} from "./importFromFileTestUtils";
 
 type ImportOptions = Raven.Client.Documents.Smuggler.DatabaseSmugglerImportOptions;
-type DatabaseItemType = Raven.Client.Documents.Smuggler.DatabaseItemType;
 type DatabaseRecordItemType = Raven.Client.Documents.Smuggler.DatabaseRecordItemType;
 type LicenseType = Raven.Server.Commercial.LicenseType;
 type Rendered = ReturnType<typeof rtlRender>;
@@ -22,85 +31,6 @@ const selectors = {
     licenseAlert: /Some data may not be imported/,
     accessAlert: /Some data cannot be imported/,
 };
-
-const adminDefaultOperateOnTypes: DatabaseItemType[] = [
-    "DatabaseRecord",
-    "Documents",
-    "Conflicts",
-    "Indexes",
-    "RevisionDocuments",
-    "Identities",
-    "CompareExchange",
-    "CounterGroups",
-    "Attachments",
-    "TimeSeries",
-    "TimeSeriesDeletedRanges",
-    "Subscriptions",
-    "Tombstones",
-    "CompareExchangeTombstones",
-];
-
-const allSettingTokens: DatabaseRecordItemType[] = [
-    "Settings",
-    "ConflictSolverConfig",
-    "Client",
-    "Revisions",
-    "Refresh",
-    "Expiration",
-    "DocumentsCompression",
-    "SchemaValidation",
-    "DataArchival",
-    "TimeSeries",
-    "Sorters",
-    "Analyzers",
-    "PostgreSQLIntegration",
-];
-
-const allOngoingTaskTokens: DatabaseRecordItemType[] = [
-    "PeriodicBackups",
-    "ExternalReplications",
-    "RavenEtls",
-    "SqlEtls",
-    "SnowflakeEtls",
-    "OlapEtls",
-    "ElasticSearchEtls",
-    "QueueEtls",
-    "HubPullReplications",
-    "SinkPullReplications",
-    "EmbeddingsGenerations",
-    "GenAiEtls",
-    "CdcSinks",
-    "AiAgents",
-    "RemoteAttachments",
-];
-
-const allConnectionStringTokens: DatabaseRecordItemType[] = [
-    "RavenConnectionStrings",
-    "SqlConnectionStrings",
-    "SnowflakeConnectionStrings",
-    "OlapConnectionStrings",
-    "ElasticSearchConnectionStrings",
-    "QueueConnectionStrings",
-    "AiConnectionStrings",
-];
-
-const without = <T,>(list: T[], ...excluded: T[]) => list.filter((x) => !excluded.includes(x));
-
-const operateOnTypesOf = (dto: ImportOptions) => dto.OperateOnTypes.split(",");
-const recordTypesOf = (dto: ImportOptions) => dto.OperateOnDatabaseRecordTypes.split(",");
-
-const adminDefaultDto: ImportOptions = {
-    IncludeExpired: true,
-    IncludeArtificial: false,
-    IncludeArchived: true,
-    TransformScript: "",
-    RemoveAnalyzers: false,
-    EncryptionKey: undefined,
-    OperateOnTypes: adminDefaultOperateOnTypes.join(",") as DatabaseItemType,
-    OperateOnDatabaseRecordTypes: "None",
-    Collections: null,
-    MaxReadOpsPerSecond: null,
-} as ImportOptions;
 
 const enterpriseAiOnlyFlags: Partial<LicenseStatus> = {
     HasGenAi: false,
